tests/wpt/mozilla/tests/mozilla/createElement_script.html

<!DOCTYPE html>
<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    var output;
    test(function() {
      output = "start";

      var script = document.createElement("script");
      script.setAttribute('type','text/javascript');
      script.textContent = "output += ' middle ';";
      document.body.appendChild(script);

      output += "end";
      assert_equals(output, "start middle end");
    });
    </script>
  </body>
</html>
